requirements.txt
Django==3.2.12
djangorestframework==3.13
psycopg2-binary
Pillow>=6.2.1,<=6.2.2

flake8==4.0.1
